Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/svg/animations/repeatn-remove-add-animation-expected.txt


FAIL Remove and add an animation element while the animation is repeating assert_equals: Expected repeatEvent event, but got beginEvent event instead expected "repeatEvent" but got "beginEvent"
